+++ recipe/passwordless/api/consumeCode.go
@@ -119,10 +119,7 @@ func ConsumeCode(apiImplementation plessmodels.APIInterface, options plessmodels
 			"status": "RESTART_FLOW_ERROR",
 		}
 	} else {
-		result = map[string]interface{}{
-			"status":  "GENERAL_ERROR",
-			"message": response.GeneralError.Message,
-		}
+		result = supertokens.ConvertGeneralErrorToJsonResponse(*response.GeneralError)
 	}
 
 	return supertokens.Send200Response(options.Res, result)