+++ recipe/passwordless/api/doesEmailExist.go
@@ -33,8 +33,12 @@ func DoesEmailExist(apiImplementation plessmodels.APIInterface, options plessmod
 	if err != nil {
 		return err
 	}
-	return supertokens.Send200Response(options.Res, map[string]interface{}{
-		"status": "OK",
-		"exists": result.OK.Exists,
-	})
+	if result.OK != nil {
+		return supertokens.Send200Response(options.Res, map[string]interface{}{
+			"status": "OK",
+			"exists": result.OK.Exists,
+		})
+	} else {
+		return supertokens.Send200Response(options.Res, supertokens.ConvertGeneralErrorToJsonResponse(*result.GeneralError))
+	}
 }